Refactor overflow menus to shared AppOverflowMenu component
